# dada-api/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchPage": {
        "description": "Cursor-paginated batch collection.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchPage",
        "type": "object"
      },
      "BatchPolicyUpdate": {
        "description": "Replacement policy for a batch that has not started.\n\nThere is no ``version`` here: a batch's policy is a snapshot rather than a\nshared resource, and the batch status is what decides whether it may still\nchange.",
        "properties": {
          "annotator_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Annotator Ids",
            "type": "array"
          },
          "mode": {
            "enum": [
              "single",
              "consensus"
            ],
            "title": "Mode",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Parameters",
            "type": "object"
          },
          "resolver": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver"
          },
          "review_thresholds": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Review Thresholds",
            "type": "object"
          }
        },
        "required": [
          "mode"
        ],
        "title": "BatchPolicyUpdate",
        "type": "object"
      },
      "BatchResponse": {
        "description": "Batch representation carrying its policy snapshot and progress.\n\nImage counts and assignment counts are reported separately because one\nimage carries one assignment per configured annotator, so the two never\ncoincide in consensus mode.",
        "properties": {
          "annotator_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Annotator Ids",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "enum": [
              "single",
              "consensus"
            ],
            "title": "Mode",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Parameters",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "initial_training",
              "test",
              "acquisition"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "requested_size": {
            "title": "Requested Size",
            "type": "integer"
          },
          "resolver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver"
          },
          "resolver_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver Version"
          },
          "review_thresholds": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Review Thresholds",
            "type": "object"
          },
          "selection_input_fingerprint": {
            "title": "Selection Input Fingerprint",
            "type": "string"
          },
          "selection_seed": {
            "title": "Selection Seed",
            "type": "integer"
          },
          "selection_strategy": {
            "title": "Selection Strategy",
            "type": "string"
          },
          "source_policy_version": {
            "title": "Source Policy Version",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "preparing",
              "annotating",
              "resolving",
              "review_required",
              "resolved",
              "closed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_assignments": {
            "title": "Submitted Assignments",
            "type": "integer"
          },
          "total_assignments": {
            "title": "Total Assignments",
            "type": "integer"
          },
          "total_items": {
            "title": "Total Items",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "purpose",
          "status",
          "mode",
          "annotator_ids",
          "resolver",
          "resolver_version",
          "parameters",
          "review_thresholds",
          "source_policy_version",
          "selection_strategy",
          "selection_seed",
          "selection_input_fingerprint",
          "requested_size",
          "total_items",
          "total_assignments",
          "submitted_assignments",
          "started_at",
          "created_at",
          "updated_at"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Stable machine- and human-readable error details.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Envelope used for all API application errors.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DADA API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/batches": {
      "get": {
        "description": "List a project's annotation batches.\n\nArgs:\n    cursor: Opaque cursor from a previous page.\n    project: Project resolved and authorized by the dependency.\n    session: Active database session.\n\nReturns:\n    One page of batches with the cursor for the next one.",
        "operationId": "list_batches_api_v1_projects__project_id__batches_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State or version conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Semantic or schema validation failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Temporarily unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Batches",
        "tags": [
          "batches"
        ]
      }
    },
    "/api/v1/projects/{project_id}/batches/{batch_id}": {
      "get": {
        "description": "Return one batch with its policy snapshot and progress.\n\nArgs:\n    batch_id: Batch being read.\n    project: Project resolved and authorized by the dependency.\n    session: Active database session.\n\nReturns:\n    The batch representation.",
        "operationId": "read_batch_api_v1_projects__project_id__batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State or version conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Semantic or schema validation failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Temporarily unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Batch",
        "tags": [
          "batches"
        ]
      },
      "patch": {
        "description": "Replace a batch's policy snapshot while it is still preparing.\n\nArgs:\n    batch_id: Batch being changed.\n    request: Validated replacement policy.\n    project: Project resolved and authorized by the dependency.\n    user: Authenticated user recorded as the actor.\n    session: Active database session.\n\nReturns:\n    The updated batch representation.",
        "operationId": "update_batch_policy_api_v1_projects__project_id__batches__batch_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State or version conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Semantic or schema validation failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Temporarily unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Batch Policy",
        "tags": [
          "batches"
        ]
      }
    },
    "/api/v1/projects/{project_id}/batches/{batch_id}/start": {
      "post": {
        "description": "Freeze a batch's policy and generate its assignments.\n\nArgs:\n    batch_id: Batch being started.\n    project: Project resolved and authorized by the dependency.\n    user: Authenticated user recorded as the actor.\n    session: Active database session.\n\nReturns:\n    The started batch representation.",
        "operationId": "start_batch_api_v1_projects__project_id__batches__batch_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State or version conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Semantic or schema validation failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Temporarily unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Batch",
        "tags": [
          "batches"
        ]
      }
    },
...
  }
}